Refactor process polling code into common utility
